# app/docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.RegisterAccountRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  controller.RegisterAccountResponse:
    properties:
      token:
        type: string
      username:
        type: string
    required:
    - token
    - username
    type: object
  controller.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  controller.SignInResponse:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  controller.SimpleResponse:
    properties:
      message:
        example: OK
        type: string
    type: object
  utils.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: ryan.murulo@gmail.com
    name: API Support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Roommates API
  version: "1.0"
paths:
  /api/v1/auth/register-account:
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: Info used to register account
        in: body
        name: RegisterAccount
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.RegisterAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: RegisterAccount new account
      tags:
      - auth
  /api/v1/auth/sign-in:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email and password
      parameters:
      - description: Information used for sign-in
        in: body
        name: SignIn
        required: true
        schema:
          $ref: '#/definitions/controller.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SignInResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: User login
      tags:
      - auth
  /api/v1/auth/sign-out:
    get:
      consumes:
      - application/json
      description: Authenticates a user with email and password
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: User login
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    description: Used for authentication of most of the access points
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"